% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augdata.R
\name{augdata}
\alias{augdata}
\alias{augdata.data.frame}
\alias{augdata.hyperframe}
\title{Augment \link[spatstat.geom]{hyperframe} for \link[mgcv]{gam}}
\usage{
augdata(data, formula, ...)

\method{augdata}{data.frame}(data, formula, ...)

\method{augdata}{hyperframe}(data, formula, ...)
}
\arguments{
\item{data}{\link[spatstat.geom]{hyperframe}}

\item{formula}{one-sided \link[stats]{formula}}

\item{...}{additional parameters, currently of no use}
}
\description{
Augment \link[spatstat.geom]{hyperframe} for \link[mgcv]{gam}
}
\author{
Tingting Zhan, Erjia Cui
}
