% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_xy.R
\name{cor_xy}
\alias{cor_xy}
\alias{cor_xy.hyper_gam}
\title{\eqn{xy}-Correlation}
\usage{
cor_xy(object, ...)

\method{cor_xy}{hyper_gam}(object, probs = 0.5, ...)
}
\arguments{
\item{object}{a \link{hyper_gam} object}

\item{...}{additional parameters, currently not in use}

\item{probs}{\link[base]{numeric} scalar or \link[base]{vector} \eqn{\tilde{p}},
taking values between 0 and 1, see function \link[stats]{quantile}.}
}
\value{
The \code{S3} method \code{\link[=cor_xy.hyper_gam]{cor_xy.hyper_gam()}} returns a \link[base]{numeric} scalar or \link[base]{vector} of
\link[stats]{cor}relation(s).
}
\description{
\eqn{xy}-Correlation
}
\details{
..
}
\keyword{internal}
