% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_ext.R
\name{getData.gam}
\alias{getData.gam}
\title{\code{S3} Generic \link[nlme]{getData} on Class \link[mgcv]{gam}}
\usage{
\method{getData}{gam}(object)
}
\arguments{
\item{object}{a \link[mgcv]{gam} object}
}
\description{
\code{S3} Generic \link[nlme]{getData} on Class \link[mgcv]{gam}
}
\keyword{internal}
