% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{hyper.gam-package}
\alias{hyper.gam}
\alias{hyper.gam-package}
\title{hyper.gam: Generalized Additive Models with Hyper Column}
\description{
An interactive HTML widget of the perspective plot for generalized additive models; an alternative solution of the function mgcv::vis.gam(). This R package author has retired from academic research. Accordingly, this package should not be considered a validated tool for use in peer-reviewed publications or as the basis for grant applications. Backward compatibility with user-code published in \doi{10.1093/bioinformatics/btaf182} and \doi{10.1093/bioinformatics/btaf430} is not maintained in versions >= 0.3.0 (June 2026) of this package. The authors of those publications are the appropriate contacts for reproducibility inquiries.
}
\note{
This R package author has retired from academic research.

Accordingly, this package should not be considered a validated tool for use in peer-reviewed publications or as the basis for grant applications.

Backward compatibility with user-code published in the following publications is not maintained in versions \verb{>= 0.3.0} of this package. The authors of those publications are the appropriate contacts for reproducibility inquiries.

\doi{10.1093/bioinformatics/btaf182}

\doi{10.1093/bioinformatics/btaf430}
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com}

Authors:
\itemize{
  \item Tingting Zhan \email{tingtingzhan@gmail.com}
}

Other contributors:
\itemize{
  \item Erjia Cui \email{ecui@umn.edu} (\href{https://orcid.org/0000-0003-3576-2892}{ORCID}) (up to v0.2.3) [contributor]
}

}
