% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_k_fold.R
\name{predict_k_fold.hyper_gam}
\alias{predict_k_fold.hyper_gam}
\title{Predict via Cross Validation}
\usage{
predict_k_fold.hyper_gam(
  object,
  k,
  mc.cores = switch(.Platform$OS.type, windows = 1L, detectCores()),
  ...
)
}
\arguments{
\item{object}{a \link{hyper_gam} object}

\item{k}{\link[base]{integer} scalar}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=predict_k_fold.hyper_gam]{predict_k_fold.hyper_gam()}} returns a \link[base]{list}.
}
\description{
Predict via Cross Validation
}
\keyword{internal}
