% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_adjust.R
\name{sign_adjust}
\alias{sign_adjust}
\alias{sign_adjust.hyper_gam}
\title{Sign Adjustment}
\usage{
sign_adjust(object, ...)

\method{sign_adjust}{hyper_gam}(object, ...)
}
\arguments{
\item{object}{..}

\item{...}{parameters of function \code{\link[=cor_xy.hyper_gam]{cor_xy.hyper_gam()}}}
}
\value{
The \code{S3} method \code{\link[=sign_adjust.hyper_gam]{sign_adjust.hyper_gam()}} returns a \link[base]{numeric} \link[base]{vector}.
}
\description{
Sign Adjustment
}
\keyword{internal}
