% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_gam.R
\name{trace_gam}
\alias{trace_gam}
\title{Trace on \link[mgcv]{gam} Model}
\usage{
trace_gam(
  object,
  formula,
  newdata = data,
  newid = seq_len(min(3L, nrow(newdata))),
  ...
)
}
\arguments{
\item{object}{a \link[mgcv]{gam} model}

\item{formula}{one-sided \link[stats]{formula}}

\item{newdata}{\link[base]{data.frame}, the test set}

\item{newid}{\link[base]{integer} scalar or \link[base]{vector},
row indices of \code{newdata} to be visualized.
Default \code{1:3}, i.e., the first three test subjects.}

\item{...}{additional parameters, to be passed into
the function \link[plotly]{plot_ly}}
}
\value{
The function \code{\link[=trace_gam]{trace_gam()}} returns a pretty \CRANpkg{htmlwidgets} created by \strong{R} package \CRANpkg{plotly}.
}
\description{
Trace on the surface of a \link[mgcv]{gam} model.
}
\note{
The maintainer is not aware of any functionality of projection of arbitrary \strong{curves} in package \CRANpkg{plotly}.
}
\examples{
library(mgcv)
# ?s
# ?ti

}
\keyword{internal}
