% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.hyper_gam}
\alias{update.hyper_gam}
\title{\link[stats]{update} a \link{hyper_gam} model}
\usage{
\method{update}{hyper_gam}(object, ...)
}
\arguments{
\item{object}{a \link{hyper_gam} model}

\item{...}{additional parameters, currently not in use}
}
\value{
The \code{S3} method \code{\link[=update.hyper_gam]{update.hyper_gam()}} returns a \link{hyper_gam} model
}
\description{
\link[stats]{update} a \link{hyper_gam} model
}
\keyword{internal}
