% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_gam.R
\name{widget_gam}
\alias{widget_gam}
\title{Alternative \link[graphics]{persp}ective Plot for \link[mgcv]{gam} Model}
\usage{
widget_gam(x, ..., colorscale)
}
\arguments{
\item{x}{a \link[mgcv]{gam} model}

\item{...}{parameters of the function \link[mgcv]{vis.gam}}

\item{colorscale}{to be passed into the function \link[plotly]{add_surface}}
}
\value{
The function \code{\link[=widget_gam]{widget_gam()}} returns a pretty \CRANpkg{htmlwidgets} created by \strong{R} package \CRANpkg{plotly}.
}
\description{
An interactive \link[graphics]{persp}ective plot for
\link[mgcv]{gam} model,
rendered as an \CRANpkg{htmlwidgets}
using the package \CRANpkg{plotly}.
}
\note{
The internal utility functions \code{persp_gam_int()} and \code{newd_gam_int()} are based on the function \link[mgcv]{vis.gam}.
}
\examples{
library(mgcv)
colorscale = list(c(0, 1), c('white', 'lightgreen'))

# examples from ?mgcv::te
test1 = \(x,z,sx=0.3,sz=0.4) { 
 x = x*20
 (pi**sx*sz)*(1.2*exp(-(x-0.2)^2/sx^2-(z-0.3)^2/sz^2)+
  0.8*exp(-(x-0.7)^2/sx^2-(z-0.8)^2/sz^2))
}
n = 500
x = runif(n)/20; z = runif(n);
xs = seq(0,1,length=30)/20; zs = seq(0,1,length=30)
pr = data.frame(x=rep(xs,30),z=rep(zs,rep(30,30)))
truth = matrix(test1(pr$x,pr$z),30,30)
persp(xs,zs,truth); title("truth")
f = test1(x,z)
y = f + rnorm(n)*0.2
b1 = gam(y ~ s(x,z))
b2 = gam(y ~ te(x,z))
b3 = gam(y ~ ti(x) + ti(z) + ti(x,z))
b4 = gam(y ~ ti(x) + ti(x,z,mc=c(0,1))) ## note z constrained!
\donttest{
widget_gam(b1, colorscale = colorscale)
widget_gam(b2, colorscale = colorscale)
widget_gam(b3, colorscale = colorscale)
widget_gam(b4, colorscale = colorscale)
}
}
