% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{hyper.gam-package}
\alias{hyper.gam}
\alias{hyper.gam-package}
\title{hyper.gam: Generalized Additive Models with Hyper Column}
\description{
An interactive HTML widget of the perspective plot for generalized additive models. An alternative solution of the function mgcv::vis.gam(). The author has retired from academic research. Accordingly, this package should not be considered a validated tool for use in peer-reviewed publications or as the basis for grant applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tingtingzhan/hyper.gam}
  \item \url{https://tingtingzhan-btaf430.netlify.app}
}

}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

Authors:
\itemize{
  \item Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})
}

Other contributors:
\itemize{
  \item Erjia Cui \email{ecui@umn.edu} (\href{https://orcid.org/0000-0003-3576-2892}{ORCID}) (up to v0.2.3) [contributor]
}

}
\keyword{internal}
