% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{hyper.gam-package}
\alias{hyper.gam}
\alias{hyper.gam-package}
\title{\link[mgcv]{gam} with one-and-only-one \link[base]{matrix} Predictor}
\description{
Function \link{hyper_gam} ..

Function \link{cor_xy.gam} ..

Function \link{sign_adjust.hyper_gam} ..

Function \link{predict.hyper_gam} ..

Functions \link{persp.hyper_gam} and \link{contour.hyper_gam} ..

Function \link{integrandSurface} ..
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

Authors:
\itemize{
  \item Inna Chervoneva \email{Inna.Chervoneva@jefferson.edu} (\href{https://orcid.org/0000-0002-9104-4505}{ORCID})
}

Other contributors:
\itemize{
  \item Erjia Cui \email{ecui@umn.edu} (\href{https://orcid.org/0000-0003-3576-2892}{ORCID}) [contributor]
}

}
