% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.hyper_gam}
\alias{predict.hyper_gam}
\title{Predicted Sign-Adjusted Quantile Indices}
\usage{
\method{predict}{hyper_gam}(
  object,
  newdata = object$data,
  sign_adjusted = TRUE,
  sgn = if (sign_adjusted) sign(cor_xy(object, probs = 0.5)) else 1,
  ...
)
}
\arguments{
\item{object}{an \link{hyper_gam} object based on the training set.}

\item{newdata}{test \link[base]{data.frame}, with at least
the response \eqn{y^{\text{new}}} and
the \link[base]{double} \link[base]{matrix} of
functional predictor values \eqn{X^{\text{new}}}
of the test set, tabulated on the same \eqn{p}-grid as the training set \eqn{X}.
If missing, the training set \code{object@gam$data} will be used.}

\item{sign_adjusted}{\link[base]{logical} scalar}

\item{sgn}{...}

\item{...}{additional parameters, currently not in use.}
}
\value{
Function \link{predict.hyper_gam} returns a
\link[base]{double} \link[base]{vector},
which is the predicted sign-adjusted quantile indices on the test set.
}
\description{
To predict sign-adjusted quantile indices of a test set.
}
\details{
Function \link{predict.hyper_gam} computes
the predicted sign-adjusted quantile indices on the test set,
which is
the product of function \link[mgcv]{predict.gam} return
and the correlation sign based on training set
(see Step 3 of section \strong{Details} of function \link{hyper_gam}).
Multiplication by this sign is required to ensure
that the predicted sign-adjusted quantile indices
are positively associated with the \strong{training} functional predictor values
at the selected tabulating grid.
}
