% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_adjust.R
\name{sign_adjust}
\alias{sign_adjust}
\alias{sign_adjust.hyper_gam}
\title{Sign Adjustment}
\usage{
sign_adjust(object, ...)

\method{sign_adjust}{hyper_gam}(object, ...)
}
\arguments{
\item{object}{..}

\item{...}{parameters of function \link{cor_xy.gam}}
}
\description{
Sign Adjustment
}
